pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esp32-wifi-monitor"
version = "1.0.0"
description = "Python wrapper to monitor WiFi performance from ESP32 devices connected via USB"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ekow Taylor", email = "ekowtaylor@meta.com"}
]
keywords = ["esp32", "wifi", "performance", "monitoring", "serial", "iot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
]

[project.scripts]
esp32-wifi-monitor = "esp32_wifi.cli:main"

[project.urls]
Homepage = "https://github.com/ekowtaylor/esp32_wifi_clients"
Repository = "https://github.com/ekowtaylor/esp32_wifi_clients.git"
Issues = "https://github.com/ekowtaylor/esp32_wifi_clients/issues"

[tool.setuptools.packages.find]
include = ["esp32_wifi*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
